Add in reload error handling
When the user would use the reload command the entire framework would fail, this was due to a unhandled exception. Now when the user reloads a bad module it catches it and testing shows it will not instantiate the new module. This will allow the old module to be used till new one is fixed.